+++ crabml-core/src/tensor/cpu/buf/mod.rs
@@ -4,5 +4,6 @@ pub mod buf_f32;
 
 pub use buf::CpuTensorBuf;
 pub use buf::CpuTensorBufIter;
+pub use buf::BlockVecCompute;
 pub use buf_q8_0::BlockQ8_0;
 pub use buf_q8_0::QuantBufQ8_0;
