+++ crabml-core/src/tensor/cpu/compute.rs
@@ -1,16 +1,13 @@
 use crate::error::ErrorKind;
 use crate::error::Result;
-use crate::gguf::GGMLType;
 use crate::tensor::cpu::buf::CpuTensorBuf;
-use crate::tensor::cpu::buf::BlockQ8_0;
-use crate::tensor::cpu::validation::require_tensor_contiguous;
-use crate::tensor::cpu::validation::require_tensor_dims;
-use crate::tensor::cpu::validation::require_tensor_matmul_2d_shapes;
-use crate::tensor::cpu::validation::require_tensor_shape;
+use crate::tensor::cpu::assert::require_tensor_contiguous;
+use crate::tensor::cpu::assert::require_tensor_dims;
+use crate::tensor::cpu::assert::require_tensor_matmul_2d_shapes;
+use crate::tensor::cpu::assert::require_tensor_shape;
 use crate::tensor::CpuTensor;
 use rayon::prelude::*;
 
-use super::buf::QuantBufQ8_0;
 use super::buf::buf::BlockVecCompute;
 
 ///! arithmetic.rs contains the tensor arithmetics operations like matmul, accum, etc.
