+++ crabml-core/src/tensor/cpu/mod.rs
@@ -1,6 +1,6 @@
 pub mod buf;
 pub mod compute;
 pub mod tensor;
-pub mod validation;
+pub mod assert;
 
 pub use tensor::CpuTensor;
